Generate updated requirements.txt for examples using pip-compile
`pip-compile` is a tool to generate requirements.txt file with locked transitive dependencies from a requirements.in containing only immediate dependencies https://github.com/jazzband/pip-tools/.

This PR adds `requirements.in` files for the examples and `pip-compile`s them. This should resolve all the dependenabot "alerts" and PRs since the dependencies will be up to date. Renovate (or dependabot maybe) should be able to do this automatically in the future.